pyproject.toml
[tool.poetry]
name = "aiarena-docker"
version = "0.1.0"
description = ""
authors = ["AI Arena <staff@aiarena.net>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "*"
aiohttp-jinja2 = "*"
attrs = "*"
burnysc2 = ">=7.0.1"
certifi = "*"
future = "*"
gym = "*"
h5py = "*"
imutils = "*"
jinja2 = "*"
loguru = "*"
more-itertools = "*"
multidict = "*"
numpy = "*"
opencv-contrib-python = "*"
opencv-python = "*"
pandas = "*"
pipenv = "*"
portpicker = "*"
protobuf = "*"
psutil = "*"
pyglet = "*"
pydantic = "*"
pygame = "*"
pysc2 = "*"
pyyaml = "*"
python-utils = "*"
requests = "*"
s2clientprotocol = "*"
scikit-image = "*"
scikit-learn = "*"
scipy = "*"
six = "*"
sqlalchemy = "*"
stable-baselines3 = "*"
tensorflow-cpu = ">=2.14.1"
termcolor = "*"
torch =  {version = "*", source = "torch-cpu"}
torch-optimizer = "*"
torchvision =  {version = "*", source = "torch-cpu"}
tqdm = "*"
urllib3 = "*"
uvloop = "*"
virtualenv = "*"
virtualenv-clone = "*"
websockets = "*"
yarl = "*"

[tool.poetry.group.dev.dependencies]
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
 name = "torch-cpu"
 url = "https://download.pytorch.org/whl/cpu"
 priority = "explicit"